requirements.txt
wxPython>=4.2.0
requests>=2.28.0
cryptography>=3.4.0
python-vlc>=3.0.18121